post Api for learning path added

+++ src/main/java/com/xebia/learningmanagement/repository/DurationRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DurationRepository extends JpaRepository<Duration,Long> {
+public interface DurationRepository extends JpaRepository<Duration,Integer> {
 }
